BOMBACLATTTT RAG FINAL/requirements.txt
pymupdf==1.25.5
langchain==0.3.24
langchain-text-splitters==0.3.8
openai==1.72.0
pinecone==6.0.0
tiktoken==0.7.0
tenacity==9.0.0
python-dotenv==1.0.1
tqdm==4.67.1
streamlit==1.42.0
supabase==2.4.0
uuid==1.30
requests==2.31.0
beautifulsoup4==4.12.2
